+++ ingestion_method/default.go
@@ -2,17 +2,17 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package ingestion_methods
+package ingestion_method
 
 import (
 	"bytes"
 	_ "embed"
 )
 
-//go:embed ingestion_methods.yml
-var defaultIngestionMethodsFile []byte
+//go:embed ingestion_method.yml
+var defaultIngestionMethodFile []byte
 
 // DefaultIngestionMethods loads the default ingestion methods from the embedded file
 func DefaultIngestionMethods() IngestionMethods {
-	return MustReadIngestionMethods(bytes.NewReader(defaultIngestionMethodsFile))
+	return MustReadIngestionMethods(bytes.NewReader(defaultIngestionMethodFile))
 }
